refactor(dbc_genesis): move to sn_interfaces
