+++ sn_node/src/node/flow_ctrl/tests/dbc_utils.rs
@@ -1,10 +1,17 @@
-use crate::node::api::gen_genesis_dbc;
+// Copyright 2022 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
 use eyre::{eyre, Result};
 use sn_dbc::{
     Dbc, KeyImage, Owner, OwnerOnce, RingCtTransaction, SpentProof, SpentProofShare, Token,
     TransactionBuilder,
 };
-use sn_interface::{messaging::data::RegisterCmd, types::ReplicatedData};
+use sn_interface::{dbcs::gen_genesis_dbc, messaging::data::RegisterCmd, types::ReplicatedData};
 use std::collections::BTreeSet;
 
 /// Get the spent proof share that's packaged inside the data that's to be replicated to the adults
